Use menu windows and allow sys mode
The menu is placed in a window with a title.

The -s option allows to use the /sys/class/gpio interface which
does not require root privileges.
However the pins must be exported prior to running this application
using gpio export.
Two shell scripts are included to export and unexport the required
pins.